change slicesdlg to selectionsdlg
