+++ nxsconfigtool/DimensionsDlg.py
@@ -178,7 +178,7 @@ class DimensionsDlg(QDialog):
 
 if __name__ == "__main__":
     import sys
-    from PyQt5.QtGui import QApplication
+    from PyQt5.QtWidgets import QApplication
 
     logging.basicConfig(level=logging.DEBUG)
