+++ nxsconfigtool/ui/sourceviewdlg.ui
@@ -44,7 +44,7 @@
            <item>
             <widget class="QLabel" name="dimLabel">
              <property name="text">
-              <string> []              </string>
+              <string>[]</string>
              </property>
             </widget>
            </item>
@@ -66,52 +66,21 @@
          <item>
           <layout class="QHBoxLayout" name="horizontalLayout_11">
            <item>
-            <widget class="QPushButton" name="slicePushButton">
+            <widget class="QPushButton" name="selectionsPushButton">
              <property name="text">
-              <string>Slices:</string>
+              <string>Selections:</string>
              </property>
             </widget>
            </item>
            <item>
-            <widget class="QLabel" name="sliceLabel">
-             <property name="text">
-              <string>[]</string>
-             </property>
-            </widget>
-           </item>
-           <item>
-            <spacer name="sliceHorizontalSpacer">
-             <property name="orientation">
-              <enum>Qt::Horizontal</enum>
-             </property>
-             <property name="sizeHint" stdset="0">
-              <size>
-               <width>40</width>
-               <height>20</height>
-              </size>
-             </property>
-            </spacer>
-           </item>
-          </layout>
-         </item>
-         <item>
-          <layout class="QHBoxLayout" name="horizontalLayout">
-           <item>
-            <widget class="QPushButton" name="slabPushButton">
-             <property name="text">
-              <string>Slabs:</string>
-             </property>
-            </widget>
-           </item>
-           <item>
-            <widget class="QLabel" name="slabLabel">
+            <widget class="QLabel" name="selectionsLabel">
              <property name="text">
               <string>[]</string>
              </property>
             </widget>
            </item>
            <item>
-            <spacer name="slabHorizontalSpacer">
+            <spacer name="selectionsHorizontalSpacer">
              <property name="orientation">
               <enum>Qt::Horizontal</enum>
              </property>
@@ -214,7 +183,7 @@
  </widget>
  <tabstops>
   <tabstop>dimPushButton</tabstop>
-  <tabstop>slicePushButton</tabstop>
+  <tabstop>selectionsPushButton</tabstop>
   <tabstop>docTextEdit</tabstop>
   <tabstop>applyPushButton</tabstop>
   <tabstop>resetPushButton</tabstop>
